project/rosa_agent/EnhanceROSAMilestone/requirements.txt
langchain
langchain-openai
langchain-community
python-dotenv
nats-py
streamlit
pydantic-ai
pydantic>=2.0
openai>=1.0




